Revert "Duration markov"
