--- scripts/convert_csv.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-
-for FILE in $(ls ${1}*.csv);
-do 
-  gsed 's/,/*/2g' $FILE | gsed 's/duration.\+/duration/g' > ${FILE}_clean
-done
